refactor: cleanup SchemaObject and its serializer
BREAKING CHANGE: SchemaObject codecs were changed
