--- src/schema/2.0/schema-object/all-of-schema-object.ts
@@ -1,8 +1,0 @@
-import { Option } from 'fp-ts/lib/Option';
-import { SchemaObject } from './schema-object';
-
-export interface AllOfSchemaObject {
-	readonly allOf: SchemaObject[];
-	readonly description: Option<string>;
-	readonly type: undefined;
-}
