--- src/schema/2.0/schema-object/array-schema-object.ts
@@ -1,6 +1,0 @@
-import { SchemaObject } from './schema-object';
-
-export interface ArraySchemaObject {
-	readonly type: 'array';
-	readonly items: SchemaObject;
-}
