--- src/schema/2.0/schema-object/reference-schema-object.ts
@@ -1,5 +1,0 @@
-import { ReferenceObject } from '../reference-object';
-
-export interface ReferenceSchemaObject extends ReferenceObject {
-	readonly type: undefined;
-}
